requirements.txt
circuitbreaker
django
djangorestframework
psycopg2-binary
requests
drf-yasg
pytest-django
pytest-cov
faker